+++ src/common/funcs.ts
@@ -8,7 +8,7 @@ dayjs.extend(isToday);
 import getLangFunc from "../lang";
 
 
-export function print(msg) {
+export function print(msg: string) {
   console.info(`#${logseq.baseInfo.id}: ${msg}`);
 }
 
@@ -476,7 +476,7 @@ export function provideStyle(opts: any = {}) {
       border-right: 1px solid #eee;
     }
 
-    [id^="logseq-plugin-block-calendar--block-calendar-yearly-slot"] .inline-button  {
+    [id^="logseq-plugin-block-calendar--block-calendar-yearly-slot"] .header .inline-button  {
       display: inline-block;
       height: 2.6rem;
     }
@@ -488,10 +488,16 @@ export function provideStyle(opts: any = {}) {
     [id^="logseq-plugin-block-calendar--block-calendar-yearly-slot"] .controls  {
       display: flex;
     }
-    [id^="logseq-plugin-block-calendar--block-calendar-yearly-slot"] .calendar-title  {
+    [id^="logseq-plugin-block-calendar--block-calendar-yearly-slot"] .calendar-header-title  {
       font-weight: bold;
       font-size: 1.5em;
       height: 60px;
+    }
+
+    [id^="logseq-plugin-block-calendar--block-calendar-yearly-slot"] .logseq-block-calendar .calendar-title  {
+      font-weight: bold;
+      font-size: 1em;
+      height: 40px;
     }
 
     [id^="logseq-plugin-block-calendar--block-calendar-yearly-slot"] .yearly-months {